requirements.txt
Pillow==9.2.0
opencv-python-headless==4.6.0.66
matplotlib==3.7.1
nba_api==1.4.1
numpy==1.24.3
pandas==1.5.3
plotly==5.9.0
rasterio==1.2.10
seaborn==0.12.2
scikit-learn==1.3.0
st_pages==0.4.5
streamlit==1.36.0
folium==0.14.0
streamlit_folium==0.18.0
